requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.2
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
openai==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
starlette==0.40.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
